remove two timetables implem
