+++ random/src/main.rs
@@ -2,7 +2,7 @@ use launch::loki::{self, DataWithIters};
 use launch::solver::Solver;
 use launch::{
     config,
-    loki::{DailyData, LoadsDailyData, LoadsPeriodicData, PeriodicData},
+    loki::{DailyData, PeriodicData},
 };
 
 use loki::log;
@@ -143,8 +143,6 @@ pub fn launch(config: Config) -> Result<(), Error> {
     match config.launch_params.data_implem {
         config::DataImplem::Periodic => config_launch::<PeriodicData>(config),
         config::DataImplem::Daily => config_launch::<DailyData>(config),
-        config::DataImplem::LoadsPeriodic => config_launch::<LoadsPeriodicData>(config),
-        config::DataImplem::LoadsDaily => config_launch::<LoadsDailyData>(config),
     }
 }
